pyproject.toml
[tool.poetry]
name = "kaleidoscope"
version = "0.1.0"
description = "kaleidoscope toy compiler in python"
authors = ["Li Chen <amazonli@amazon.com>"]

[tool.poetry.dependencies]
python = "^3.9"
antlr4-python3-runtime = "^4.9.2"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
pylint = "^2.8.3"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
